Basic idea
The basic idea is to move towards a more functional approach. As you can see the code is a lot tighter than before but still retains its functionality. There is one major issue still outstanding and that is the rerendering issue. I need to look into this but you can run the demo and see that it does apply the rotation as expected albeit with a rerender for each frame (which we don't want). Will keep working on this over the coming while but appreciate any feedback on the pros/cons of this implementation and also any ideas on how to resolve the rerendering issue that you may have. Until then I will see what I can do on my end and update as I go.